website/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-appconf==1.0.5
django-bootstrap5==22.2
django-compressor==4.3.1
django-cors-headers==3.13.0
django-libsass==0.9
djangorestframework==3.14.0
libsass==0.22.0
Pillow==9.4.0
pytz==2022.7.1
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
